refactor(e2e): take the receiver URL from the harness
Nine sites across eight suites spelled http://%s-test-receiver:8080/%s
inline while fixture.ReceiverURL had no caller outside tenancy.go. The
harness already owns the receiver, so the URL belongs beside MustReceive.

Receiver moves to its own file and gains URL, URLIn for a client in
another namespace, and ReceiverName, which the log selector and both
URLs now derive from instead of respelling the suffix.

The package functions take a release rather than an Env so the seven
unmigrated suites can use them too; they go when the last one moves,
leaving the methods. common's two copies stay because common cannot
import the harness, and they go with common.
