+++ e2e/internal/harness/harness.go
@@ -18,15 +18,12 @@ import (
 	"context"
 	"fmt"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"slices"
-	"strings"
 	"sync"
 	"testing"
 	"time"
 
-	"github.com/cisco-open/operator-tools/pkg/types"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	appsv1 "k8s.io/api/apps/v1"
@@ -51,10 +48,6 @@ const (
 	clusterStopTimeout = time.Minute
 
 	clusterLogLimit = 100 * 1000
-
-	// Generous on purpose: more lines makes a tag easier to find and its
-	// absence harder to claim.
-	receiverLogTail = 100
 
 	// waitBudget and waitInterval are what the suites spend by hand today.
 	waitBudget   = 5 * time.Minute
@@ -226,55 +219,6 @@ func (e *Env) WaitForRunning(conditions ...wait.Condition) {
 		}
 		return true
 	}, e.waitBudget(), waitInterval, "still waiting for %s", &outstanding)
-}
-
-// Receiver is the test receiver the chart installs, where a suite looks to see
-// which logs arrived.
-type Receiver struct {
-	env *Env
-}
-
-// MustReceive does not echo the tail each poll: the archived cluster dump
-// already carries the receiver's log.
-func (r Receiver) MustReceive(tags ...string) {
-	r.env.T.Helper()
-
-	var outstanding pending
-	require.Eventuallyf(r.env.T, func() bool {
-		logs, err := r.Logs()
-		if err != nil {
-			r.env.T.Logf("reading the test receiver: %v", err)
-			return false
-		}
-		for _, tag := range tags {
-			if !strings.Contains(logs, tag) {
-				outstanding.set(tag)
-				return false
-			}
-		}
-		return true
-	}, r.env.waitBudget(), waitInterval, "the test receiver never logged %s", &outstanding)
-}
-
-// MustNotReceive is a point-in-time check, since an absence cannot be waited
-// for. It belongs after whatever wait establishes that the pipeline is running.
-func (r Receiver) MustNotReceive(tags ...string) {
-	r.env.T.Helper()
-
-	logs, err := r.Logs()
-	require.NoError(r.env.T, err)
-	for _, tag := range tags {
-		assert.NotContains(r.env.T, logs, tag)
-	}
-}
-
-func (r Receiver) Logs() (string, error) {
-	out, err := common.CmdEnv(exec.Command("kubectl",
-		"logs",
-		"-n", r.env.ControlNamespace,
-		"--tail", fmt.Sprint(receiverLogTail),
-		"-l", fmt.Sprintf("%s=%s-test-receiver", types.NameLabel, r.env.Release)), r.env.Cluster).Output()
-	return string(out), err
 }
 
 // waitBudget holds a wait under what is left of the binary's deadline, so one
