+++ e2e/syslog-ng-aggregator-detached/syslog_ng_aggregator_detached_test.go
@@ -41,6 +41,7 @@ import (
 
 	"github.com/kube-logging/logging-operator/e2e/common"
 	"github.com/kube-logging/logging-operator/e2e/common/setup"
+	"github.com/kube-logging/logging-operator/e2e/internal/harness"
 	"github.com/kube-logging/logging-operator/e2e/internal/image"
 	"github.com/kube-logging/logging-operator/e2e/internal/wait"
 	"github.com/kube-logging/logging-operator/pkg/resources/syslogng"
@@ -215,7 +216,7 @@ func TestSyslogNGDetachedIsRunningAndForwardingLogs(t *testing.T) {
 			},
 			Spec: v1beta1.SyslogNGOutputSpec{
 				HTTP: &syslogngoutput.HTTPOutput{
-					URL: fmt.Sprintf("http://%s-test-receiver:8080/%s", releaseNameOverride, testTag),
+					URL: harness.ReceiverURL(releaseNameOverride, testTag),
 					Headers: []string{
 						"Content-type: application/json",
 					},
@@ -312,7 +313,7 @@ func TestSyslogNGDetachedIsRunningAndForwardingLogs(t *testing.T) {
 			cmd := common.CmdEnv(exec.Command("kubectl",
 				"logs",
 				"-n", ns,
-				"-l", fmt.Sprintf("%s=%s-test-receiver", types.NameLabel, releaseNameOverride)), c)
+				"-l", fmt.Sprintf("%s=%s", types.NameLabel, harness.ReceiverName(releaseNameOverride))), c)
 			rawOut, err := cmd.Output()
 			if err != nil {
 				t.Logf("failed to get log consumer logs: %+v %s", err, rawOut)
