+++ e2e/volumedrain/volumedrain_test.go
@@ -40,6 +40,7 @@ import (
 
 	"github.com/kube-logging/logging-operator/e2e/common"
 	"github.com/kube-logging/logging-operator/e2e/common/setup"
+	"github.com/kube-logging/logging-operator/e2e/internal/harness"
 	"github.com/kube-logging/logging-operator/e2e/internal/image"
 	"github.com/kube-logging/logging-operator/e2e/internal/wait"
 	"github.com/kube-logging/logging-operator/pkg/sdk/logging/api/v1beta1"
@@ -127,7 +128,7 @@ func TestVolumeDrain_Downscale(t *testing.T) {
 			},
 			Spec: v1beta1.OutputSpec{
 				HTTPOutput: &output.HTTPOutputConfig{
-					Endpoint:    fmt.Sprintf("http://%s-test-receiver:8080/%s", releaseNameOverride, testTag),
+					Endpoint:    harness.ReceiverURL(releaseNameOverride, testTag),
 					ContentType: "application/json",
 					Buffer: &output.Buffer{
 						Type:        "file",
@@ -192,7 +193,7 @@ func TestVolumeDrain_Downscale(t *testing.T) {
 			cmd := common.CmdEnv(exec.Command("kubectl",
 				"logs",
 				"-n", ns,
-				"-l", fmt.Sprintf("%s=%s-test-receiver", types.NameLabel, releaseNameOverride)), c)
+				"-l", fmt.Sprintf("%s=%s", types.NameLabel, harness.ReceiverName(releaseNameOverride))), c)
 			rawOut, err := cmd.Output()
 			if err != nil {
 				t.Logf("failed to get log consumer logs: %v", err)
@@ -203,7 +204,7 @@ func TestVolumeDrain_Downscale(t *testing.T) {
 		}, 5*time.Minute, 3*time.Second)
 
 		cmd := common.CmdEnv(exec.Command("kubectl", "scale",
-			fmt.Sprintf("deployment/%s-test-receiver", releaseNameOverride),
+			"deployment/"+harness.ReceiverName(releaseNameOverride),
 			"-n", ns,
 			"--replicas", "0"), c)
 		common.RequireNoError(t, cmd.Run())
@@ -238,7 +239,7 @@ func TestVolumeDrain_Downscale(t *testing.T) {
 		require.Eventually(t, wait.PodShouldBeRunning(t, c.GetClient(), client.ObjectKey{Namespace: ns, Name: fluentdReplicaName}), 30*time.Second, time.Second/2)
 
 		cmd = common.CmdEnv(exec.Command("kubectl", "scale",
-			fmt.Sprintf("deployment/%s-test-receiver", releaseNameOverride),
+			"deployment/"+harness.ReceiverName(releaseNameOverride),
 			"-n", ns,
 			"--replicas", "1"), c)
 		common.RequireNoError(t, cmd.Run())
@@ -357,7 +358,7 @@ func TestVolumeDrain_Downscale_DeleteVolume(t *testing.T) {
 			},
 			Spec: v1beta1.OutputSpec{
 				HTTPOutput: &output.HTTPOutputConfig{
-					Endpoint:    fmt.Sprintf("http://%s-test-receiver:8080/%s", releaseNameOverride, testTag),
+					Endpoint:    harness.ReceiverURL(releaseNameOverride, testTag),
 					ContentType: "application/json",
 					Buffer: &output.Buffer{
 						Type:        "file",
@@ -415,7 +416,7 @@ func TestVolumeDrain_Downscale_DeleteVolume(t *testing.T) {
 			cmd := common.CmdEnv(exec.Command("kubectl",
 				"logs",
 				"-n", ns,
-				"-l", fmt.Sprintf("%s=%s-test-receiver", types.NameLabel, releaseNameOverride)), c)
+				"-l", fmt.Sprintf("%s=%s", types.NameLabel, harness.ReceiverName(releaseNameOverride))), c)
 			rawOut, err := cmd.Output()
 			if err != nil {
 				t.Logf("failed to get log consumer logs: %v", err)
@@ -426,7 +427,7 @@ func TestVolumeDrain_Downscale_DeleteVolume(t *testing.T) {
 		}, 5*time.Minute, 2*time.Second)
 
 		cmd := common.CmdEnv(exec.Command("kubectl", "scale",
-			fmt.Sprintf("deployment/%s-test-receiver", releaseNameOverride),
+			"deployment/"+harness.ReceiverName(releaseNameOverride),
 			"-n", ns,
 			"--replicas", "0"), c)
 		common.RequireNoError(t, cmd.Run())
@@ -455,7 +456,7 @@ func TestVolumeDrain_Downscale_DeleteVolume(t *testing.T) {
 		require.Eventually(t, wait.PodShouldBeRunning(t, c.GetClient(), client.ObjectKey{Namespace: ns, Name: fluentdReplicaName}), 30*time.Second, time.Second/2)
 
 		cmd = common.CmdEnv(exec.Command("kubectl", "scale",
-			fmt.Sprintf("deployment/%s-test-receiver", releaseNameOverride),
+			"deployment/"+harness.ReceiverName(releaseNameOverride),
 			"-n", ns,
 			"--replicas", "1"), c)
 		common.RequireNoError(t, cmd.Run())
